% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import.R
\name{from}
\alias{from}
\title{Generate import::from calls}
\usage{
from(.pkgs = NULL, .output = c("clipboard", "cat", "return"),
  .comment = TRUE)
}
\arguments{
\item{.pkgs}{\code{[character()} or \code{list()]}

  Packages to generate import calls for.}

\item{.output}{\code{[character(1)]}

  What should happen with the generated calls? If \code{"clipboard"} (default),
  copy to clipboard. If \code{"cat"}, print to the console. If \code{"return"},
  simply return as a character vector.}

\item{.comment}{\code{[logical(1)]}

  Should comments be prepended to the generated code? Default: \code{TRUE}.}
}
\value{
\code{[character()]}

  The generated output, invisibly unless \code{output} is set to \code{"return"}.
}
\description{
This function generates calls to import::\link[import]{from} that imports
all symbols in a package.
}
\examples{
from("rpart", .output = "cat")
}

