% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import-from.R
\name{importFrom}
\alias{importFrom}
\title{Generate @importFrom directives}
\usage{
importFrom(..., .pkgs = NULL, .output = c("clipboard", "cat", "return"),
  .comment = TRUE)
}
\arguments{
\item{...}{\code{[character(1)]}

  Packages to generate \code{@importFrom}
  directives for.}

\item{.pkgs}{\code{[character()} or \code{list()]}

  Packages to generate \code{@importFrom}
  directives for, as list.}

\item{.output}{\code{[character(1)]}

  What should happen with the generated directives? If \code{"clipboard"} (default),
  copy to clipboard. If \code{"cat"}, print to the console. If \code{"return"},
  simply return as a character vector.}

\item{.comment}{\code{[logical(1)]}

  Should comments be prepended to the generated code? Default: \code{TRUE}.}
}
\value{
\code{[character()]}

  The generated output, invisibly unless \code{output} is set to \code{"return"}.
}
\description{
This function generates \code{@importFrom}
roxygen directives that import all symbols in a package.
}
\examples{
importFrom("rpart", .output = "cat")
}

